Adding default implementations to Values.equals methods
The equals methods now return false by default. They are still
overridden (with the exact same implementation) to be final in
abstract subclasses, where equality can be ruled out for certain values.
This makes the concrete classes shorter, since they only need to
implement the equals method(s) which can return true.
