+++ community/values/src/main/java/org/neo4j/values/storable/CharArray.java
@@ -19,14 +19,8 @@
  */
 package org.neo4j.values.storable;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.OffsetTime;
-import java.time.ZonedDateTime;
 import java.util.Arrays;
 
-import org.neo4j.graphdb.spatial.Geometry;
 import org.neo4j.values.AnyValue;
 import org.neo4j.values.ValueMapper;
 
@@ -65,48 +59,6 @@ public class CharArray extends TextArray
     public boolean equals( String[] x )
     {
         return PrimitiveArrayValues.equals( value, x );
-    }
-
-    @Override
-    public boolean equals( Geometry[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( ZonedDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalDate[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( DurationValue[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( OffsetTime[] x )
-    {
-        return false;
     }
 
     @Override
