+++ community/values/src/main/java/org/neo4j/values/storable/DateArray.java
@@ -20,10 +20,6 @@
 package org.neo4j.values.storable;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.OffsetTime;
-import java.time.ZonedDateTime;
 import java.util.Arrays;
 
 import org.neo4j.values.ValueMapper;
@@ -57,39 +53,9 @@ public class DateArray extends TemporalArray<LocalDate, DateValue>
     }
 
     @Override
-    public boolean equals( ZonedDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
     public boolean equals( LocalDate[] x )
     {
         return Arrays.equals( value, x);
-    }
-
-    @Override
-    public boolean equals( DurationValue[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( OffsetTime[] x )
-    {
-        return false;
     }
 
     @Override
