+++ community/values/src/main/java/org/neo4j/values/storable/DateTimeArray.java
@@ -19,10 +19,6 @@
  */
 package org.neo4j.values.storable;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.OffsetTime;
 import java.time.ZonedDateTime;
 import java.util.Arrays;
 
@@ -60,36 +56,6 @@ public class DateTimeArray extends TemporalArray<ZonedDateTime, DateTimeValue>
     public boolean equals( ZonedDateTime[] x )
     {
         return Arrays.equals( value, x);
-    }
-
-    @Override
-    public boolean equals( LocalDate[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( DurationValue[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( OffsetTime[] x )
-    {
-        return false;
     }
 
     @Override
