+++ community/values/src/main/java/org/neo4j/values/storable/NumberArray.java
@@ -34,61 +34,61 @@ public abstract class NumberArray extends ArrayValue
     abstract int compareTo( FloatingPointArray other );
 
     @Override
-    public boolean equals( boolean[] x )
+    public final boolean equals( boolean[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( char[] x )
+    public final boolean equals( char[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( String[] x )
+    public final boolean equals( String[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( Geometry[] x )
+    public final boolean equals( Geometry[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( ZonedDateTime[] x )
+    public final boolean equals( ZonedDateTime[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( LocalDate[] x )
+    public final boolean equals( LocalDate[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( DurationValue[] x )
+    public final boolean equals( DurationValue[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( LocalDateTime[] x )
+    public final boolean equals( LocalDateTime[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( LocalTime[] x )
+    public final boolean equals( LocalTime[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( OffsetTime[] x )
+    public final boolean equals( OffsetTime[] x )
     {
         return false;
     }
