+++ community/values/src/main/java/org/neo4j/values/storable/PointArray.java
@@ -19,11 +19,6 @@
  */
 package org.neo4j.values.storable;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-import java.time.OffsetTime;
-import java.time.ZonedDateTime;
 import java.util.Arrays;
 
 import org.neo4j.graphdb.spatial.Geometry;
@@ -64,105 +59,9 @@ public class PointArray extends NonPrimitiveArray<PointValue>
     }
 
     @Override
-    public boolean equals( byte[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( short[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( int[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( long[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( float[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( double[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( boolean[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( char[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( String[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( ZonedDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalDate[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( DurationValue[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( OffsetTime[] x )
-    {
-        return false;
-    }
-
-    @Override
     public ValueGroup valueGroup()
     {
         return ValueGroup.GEOMETRY_ARRAY;
-    }
-
-    @Override
-    public NumberType numberType()
-    {
-        return NumberType.NO_NUMBER;
     }
 
     @Override
