+++ community/values/src/main/java/org/neo4j/values/storable/PointValue.java
@@ -69,36 +69,6 @@ public class PointValue extends ScalarValue implements Comparable<PointValue>, P
     }
 
     @Override
-    public boolean equals( boolean x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( long x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( double x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( char x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( String x )
-    {
-        return false;
-    }
-
-    @Override
     public boolean equals( Value other )
     {
         if ( other instanceof PointValue )
