+++ community/values/src/main/java/org/neo4j/values/storable/ScalarValue.java
@@ -33,97 +33,97 @@ import org.neo4j.graphdb.spatial.Geometry;
 abstract class ScalarValue extends Value
 {
     @Override
-    public boolean equals( byte[] x )
+    public final boolean equals( byte[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( short[] x )
+    public final boolean equals( short[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( int[] x )
+    public final boolean equals( int[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( long[] x )
+    public final boolean equals( long[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( float[] x )
+    public final boolean equals( float[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( double[] x )
+    public final boolean equals( double[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( boolean[] x )
+    public final boolean equals( boolean[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( char[] x )
+    public final boolean equals( char[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( String[] x )
+    public final boolean equals( String[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( Geometry[] x )
+    public final boolean equals( Geometry[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( ZonedDateTime[] x )
+    public final boolean equals( ZonedDateTime[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( LocalDate[] x )
+    public final boolean equals( LocalDate[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( DurationValue[] x )
+    public final boolean equals( DurationValue[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( LocalDateTime[] x )
+    public final boolean equals( LocalDateTime[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( LocalTime[] x )
+    public final boolean equals( LocalTime[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( OffsetTime[] x )
+    public final boolean equals( OffsetTime[] x )
     {
         return false;
     }
