+++ community/values/src/main/java/org/neo4j/values/storable/TextArray.java
@@ -19,8 +19,6 @@
  */
 package org.neo4j.values.storable;
 
-import org.neo4j.values.ValueMapper;
-
 public abstract class TextArray extends ArrayValue
 {
     public abstract String stringValue( int offset );
@@ -31,43 +29,43 @@ public abstract class TextArray extends ArrayValue
     }
 
     @Override
-    public boolean equals( byte[] x )
+    public final boolean equals( byte[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( short[] x )
+    public final boolean equals( short[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( int[] x )
+    public final boolean equals( int[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( long[] x )
+    public final boolean equals( long[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( float[] x )
+    public final boolean equals( float[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( double[] x )
+    public final boolean equals( double[] x )
     {
         return false;
     }
 
     @Override
-    public boolean equals( boolean[] x )
+    public final boolean equals( boolean[] x )
     {
         return false;
     }
