+++ community/values/src/main/java/org/neo4j/values/storable/TimeArray.java
@@ -19,11 +19,7 @@
  */
 package org.neo4j.values.storable;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.time.OffsetTime;
-import java.time.ZonedDateTime;
 import java.util.Arrays;
 
 import org.neo4j.values.ValueMapper;
@@ -60,36 +56,6 @@ public class TimeArray extends TemporalArray<OffsetTime, TimeValue>
     public boolean equals( OffsetTime[] x )
     {
         return Arrays.equals( value, x);
-    }
-
-    @Override
-    public boolean equals( LocalDate[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( DurationValue[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( ZonedDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalDateTime[] x )
-    {
-        return false;
-    }
-
-    @Override
-    public boolean equals( LocalTime[] x )
-    {
-        return false;
     }
 
     @Override
